Applied a new formatting

+++ src/dist.hpp
@@ -110,19 +110,19 @@ private:
 
   static auto compute_skips(size_t i, size_t n) -> size_t {
     constexpr size_t BITS_IN_BYTE = 8;
-    size_t skips = 0;
+    size_t           skips        = 0;
     while (i != 0 && n != 0) {
-      size_t first_index = sizeof(i) * BITS_IN_BYTE - __builtin_clzll(i | 1);
-      skips += compute_skips_power_of_2(first_index, n);
-      n -= 1;
-      i -= 1 << (first_index - 1);
+      size_t first_index  = sizeof(i) * BITS_IN_BYTE - __builtin_clzll(i | 1);
+      skips              += compute_skips_power_of_2(first_index, n);
+      n                  -= 1;
+      i                  -= 1 << (first_index - 1);
     }
     skips += i;
     return skips;
   }
 
   uint64_t _dist;
-  uint8_t _regions;
+  uint16_t _regions;
 };
 
 bool valid_dist(dist_t d, const substitution_model_t &model);
@@ -139,9 +139,9 @@ public:
   dist_t final_state;
 };
 
-transition_t sample(dist_t init_dist, const substitution_model_t &model,
+transition_t sample(dist_t                                  init_dist,
+                    const substitution_model_t             &model,
                     std::uniform_random_bit_generator auto &gen) {
-
   auto [d, e] = model.rates();
 
   bool singleton = init_dist.popcount() == 1;
@@ -152,9 +152,7 @@ transition_t sample(dist_t init_dist, const substitution_model_t &model,
   std::vector<transition_t> rolls(model.region_count());
 
   for (size_t i = 0; i < model.region_count(); ++i) {
-    if (init_dist[i] && singleton) {
-      continue;
-    }
+    if (init_dist[i] && singleton) { continue; }
     double waiting_time = init_dist[i] ? exp_die(gen) : dis_die(gen);
     rolls[i] = transition_t{waiting_time, init_dist, init_dist.negate_bit(i)};
   }
@@ -165,18 +163,17 @@ transition_t sample(dist_t init_dist, const substitution_model_t &model,
 }
 
 std::vector<transition_t>
-generate_samples(dist_t init_dist, double brlen,
-                 const substitution_model_t &model,
+generate_samples(dist_t                                  init_dist,
+                 double                                  brlen,
+                 const substitution_model_t             &model,
                  std::uniform_random_bit_generator auto &gen) {
   std::vector<transition_t> results;
   while (true) {
-    auto r = sample(init_dist, model, gen);
-    brlen -= r.waiting_time;
-    if (brlen < 0.0) {
-      return results;
-    }
-    LOG_DEBUG("adding transition from %b to %b", r.initial_state,
-              r.final_state);
+    auto r  = sample(init_dist, model, gen);
+    brlen  -= r.waiting_time;
+    if (brlen < 0.0) { return results; }
+    LOG_DEBUG(
+        "adding transition from %b to %b", r.initial_state, r.final_state);
     init_dist = r.final_state;
     results.push_back(r);
   }
@@ -193,7 +190,8 @@ generate_samples(dist_t init_dist, double brlen,
  *  occur, but the idea maps well, so I use it internally.
  */
 std::pair<dist_t, dist_t>
-split_dist(dist_t init_dist, const substitution_model_t &model,
+split_dist(dist_t                                  init_dist,
+           const substitution_model_t             &model,
            std::uniform_random_bit_generator auto &gen) {
   // Singleton case
   if (init_dist.popcount() == 1) {
@@ -220,9 +218,7 @@ split_dist(dist_t init_dist, const substitution_model_t &model,
   }
 
   std::bernoulli_distribution left_or_right(0.5);
-  if (left_or_right(gen)) {
-    std::swap(left_dist, right_dist);
-  }
+  if (left_or_right(gen)) { std::swap(left_dist, right_dist); }
 
   LOG_DEBUG("Spilt %b into: %b, %b", init_dist, left_dist, right_dist);
   return {left_dist, right_dist};
